plc/zhou/pythonconnent/仿真/requirements.txt
python-snap7>=1.3
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0
